refactor: 코드 리뷰 반영 - AcceptanceTest 추상 클래스 생성
