+++ src/test/java/nextstep/subway/AcceptanceTest.java
@@ -1,0 +1,26 @@
+package nextstep.subway;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+import io.restassured.RestAssured;
+import nextstep.subway.utils.DatabaseCleaner;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public abstract class AcceptanceTest {
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    DatabaseCleaner databaseCleaner;
+
+    @BeforeEach
+    public void setUp() {
+        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
+            RestAssured.port = port;
+        }
+        databaseCleaner.execute();
+    }
+}
